refactor(campaign-eta): relocate formatDuration to its own module (U3)
Move formatDuration out of the soon-to-be-retired campaign-eta.ts proxy
into lib/duration.ts so the per-attack ETA formatter keeps working after
U6 deletes the proxy. attack-format.ts and campaign-eta.ts both import
from duration.ts; no behavior change, no circular dependency.
